Undo most of the class changes to enforce usage of repositories without name and instead use an assert
While it made the expected class very explicit on the side of dependencies, the inheritance was becoming quite complex
This solution relies on using the correct '.withoutName' call at the right moment(s), the assert assures that it is done correctly
