+++ src/main/scala/io/viash/config/dependencies/LocalRepository.scala
@@ -40,7 +40,7 @@ case class LocalRepository(
   tag: Option[String] = None,
   path: Option[String] = None,
   localPath: String = ""
-) extends RepositoryWithoutName with LocalRepositoryTrait {
+) extends LocalRepositoryTrait {
 
   def copyRepo(
     `type`: String,
